+++ lib/views/view_payees/picker_payee.dart
@@ -1,20 +1,20 @@
 import 'package:flutter/material.dart';
 import 'package:money/storage/data/data.dart';
-import 'package:money/models/money_objects/categories/category.dart';
+import 'package:money/models/money_objects/payees/payee.dart';
 import 'package:money/widgets/picker_edit_box.dart';
 
-Widget pickerCategory({
-  required final Category? itemSelected,
-  required final Function(Category?) onSelected,
+Widget pickerForPayee({
+  required final Payee? itemSelected,
+  required final Function(Payee?) onSelected,
 }) {
-  final List<String> options = Data().categories.getListSorted().map((element) => element.name.value).toList();
+  final List<String> options = Data().payees.getListSorted().map((element) => element.name.value).toList();
   String selectedName = itemSelected == null ? '' : itemSelected.name.value;
 
   return PickerEditBox(
     options: options,
     initialValue: selectedName,
     onChanged: (String newSelection) {
-      final Category? found = Data().categories.getByName(newSelection);
+      final Payee? found = Data().payees.getByName(newSelection);
       if (found != null) {
         onSelected(found);
       }